Add files via upload
This script produces a simple but highly useful investment calculator in the form of a graphical user interface. The calculator uses a user defined monthly investment rate, interest rate and number of years. The script uses the tkinter library to produce the GUI and the calculator is a step up from a console based investment calculator produced a couple of months ago

Libaries used:
- Tkinter
- random